Separate McpDmlToolsOption class from McpRuntimeOptions
